step 14 to 16
I am not sure how to make it work it wont update

Select the h1 with the hero section
Set width to 75%
Add padding to all sides with value .5em
Center the content using the property margin without applying specific values. Think automatically calculated on left and right.
Change text color to white
Align text in the center
Change font size to 2rem
Apply a text shadow with offsets and blur of 2px, and set color using rgba(0,0,0,.8)
Change background color to use rgba(0,0,0,.25)
Select the section with the class of trends
Change background color to purple
Select the figure element within the section with the class of trends
Set width to 22%
Apply an appropriate display property value so that you are able to display the figures beside one another
Align the figures to be centered within its parent container